% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{covr}
\alias{covr}
\alias{report}
\alias{package_coverage}
\title{'covr' overrides}
\usage{
report(
  x = dir::package_coverage(),
  file = file.path(tempdir(), paste0(attr(x, "package")$package \%||\% "coverage",
    "-report.html")),
  browse = interactive()
)

package_coverage(
  path = ".",
  type = c("tests", "vignettes", "examples", "all", "none"),
  combine_types = TRUE,
  relative_path = TRUE,
  quiet = TRUE,
  clean = TRUE,
  line_exclusions = NULL,
  function_exclusions = NULL,
  code = character(),
  install_path = normalizePath(tempfile("R_LIBS"), mustWork = FALSE),
  ...,
  exclusions,
  pre_clean = TRUE
)
}
\arguments{
\item{x}{a coverage dataset, defaults to running \code{package_coverage()}.}

\item{file}{The report filename.}

\item{browse}{whether to open a browser to view the report.}
}
\value{
These functions return \code{NULL} invisibly, they're called for side effects.
}
\description{
Wrappers around 'covr' functions that behave the same but support a flexible
folder structure. Not needed if you use \code{use_dir_package(, patch = TRUE)},
which will patch the actual 'covr' functions directly.
}
\seealso{
\code{\link[devtools:document]{devtools::document()}}, \code{\link[devtools:check]{devtools::check()}}, \code{\link[devtools:build]{devtools::build()}}
}
