% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\alias{create.character}
\alias{create.list}
\title{Create folders if they do not exist}
\usage{
create(paths, ...)

\method{create}{character}(paths, ...)

\method{create}{list}(paths, ...)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}

\item{...}{additional arguments to pass to \code{\link{dir.create}}.}
}
\value{
\code{data.table} of \code{paths} and whether or not they were created,
 invisibly.
}
\description{
A wrapper around \code{\link{dir.create}} that creates multiple folders 
   if they do not already exist.
}
\details{
Rather than \code{\link{dir.create}}, which by default tries to
 create the directory regardless of its presence, \code{create} only
 runs \code{\link{dir.create}} for folders whose \code{paths} do not 
 already exist.
}
\examples{
 \donttest{
  create("./new_folder")
 }

}
