% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_dir}
\alias{create_dir}
\title{Flexibly create a directory}
\usage{
create_dir(tree, quiet = FALSE, ...)
}
\arguments{
\item{tree}{}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{...}{additional arguments to pass to \code{\link{dir.create}}.}
}
\value{
f
}
\description{
Creates the specified folder structure
 for a directory as well as a YAML file that tracks the
 setup configurations.
}
\details{
At each level of the hierarchy, the more
 basal folders are required to be present (verified using 
 \code{\link{verify}}) and the relevant folders at that level are created 
 if not present (using \code{\link{create}}).
}
\examples{
 \donttest{
  create_dir()
 }

}
