% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{depth}
\alias{depth}
\title{Determine the depth of a list}
\usage{
depth(x)
}
\arguments{
\item{x}{Focal \code{list} to be evaluated.}
}
\value{
\code{integer} value of the depth of the list.
}
\description{
Measure the depth a list's nesting.
}
\details{
If \code{xlist = list()}, then technically the input value is a 
 list, but is empty (of length \code{0}), so depth is returned as \code{0}.
}
\examples{
 list1 <- list(main  = "main", 
               other = "other")
 list2 <- list(main  = list(data    = "data", 
                            code    = "code",
                            results = "results"), 
               other = "other")
 list3 <- list(main  = list(data    = "data", 
                            code    = "code",
                            results = list("2021-01-01" = "2021-01-01",
                                           "2021-01-02" = "2021-01-02")), 
               other = "other")
 
 depth(list1)
 depth(list2)
 depth(list3)

}
