% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_package.R
\docType{package}
\name{dir}
\alias{dir}
\title{Cross-Platform Directory Creation and Management Utilities}
\description{
Create and manage directories for R analyses projects. This 
 package particularly is designed for supporting automated analyses 
 using minimal dependencies.
}
\keyword{package}
