% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directree.R
\name{dirs.create}
\alias{dirs.create}
\title{Flexibly Create Multiple Directories}
\usage{
dirs.create(paths, showWarnings = TRUE, recursive = FALSE, mode = "0777")
}
\arguments{
\item{paths}{\code{character} vector containing multiple path names. 
See \code{\link[base]{dir.create}}.}

\item{showWarnings}{\code{logical}: should the warnings on failure be 
shown? 
See \code{\link[base]{dir.create}}.}

\item{recursive}{\code{logical}: should elements of the path other than 
the last be created? If \code{true}, like the Unix command 
\code{mkdir -p}.
See \code{\link[base]{dir.create}}.}

\item{mode}{The mode to be used on Unix-alikes: it will be coerced by 
\code{\link[base]{as.octmode}}. For \code{\link[base]{Sys.chmod}} it 
is recycled along paths.
See \code{\link[base]{dir.create}}.}
}
\value{
\code{logical} vector of results of \code{\link[base]{dir.create}}
 on each of the \code{paths} (which name the vector).
 See \code{\link[base]{dir.create}}.
}
\description{
Creates multiple directories at once by running
 \code{\link[base]{dir.create}} over multiple 
 \code{paths} through \code{\link[base]{mapply}}.
}
\examples{
 \dontrun{

   paths <- normalized_file_path(".", c("folder1", "folder2"),
                                 mustWork = FALSE)
   dirs.create(paths)
 
 }

}
