% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{finder}
\alias{finder}
\title{Directory listing}
\usage{
finder(x = ".", class = NULL, max = Inf, show_hidden = FALSE, open = TRUE)
}
\arguments{
\item{x}{A path or a finder object.}

\item{class}{An optional subclass to add.}
}
\description{
Create a HTML directory listing that is best used
for R Markdown documents with HTML outputs, including
for HTML presentations.
}
\details{
\dontrun{
finder(system.file(package = "dir"))
}
}
