% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file}
\alias{is_file}
\title{Test if given path is a file or not.}
\usage{
is_file(path)
}
\arguments{
\item{x}{A vector of string}
}
\value{
Returns a logical value indicating whether it is a file or not.
}
\description{
Test if given path is a file or not.
}
