% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_listing}
\alias{is_listing}
\title{Test if given object is finder class}
\usage{
is_listing(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value.
}
\description{
Test if given object is finder class
}
