% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{listing}
\alias{listing}
\title{Directory listing}
\usage{
listing(
  path = ".",
  recurse = TRUE,
  show_hidden = FALSE,
  open = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{path}{A path.}

\item{recurse}{If \code{TRUE} recurse fully, if a positive number the number of levels to recurse.}

\item{show_hidden}{If \code{TRUE} show hidden files and folders.}

\item{open}{Whether the folders should be open by default.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{The element ID.}
}
\value{
A list of files and folders.
}
\description{
Create a HTML directory listing that is best used
for R Markdown documents with HTML outputs, including
for HTML presentations.
}
\examples{
\dontrun{
listing(system.file(package = "dir"))
}

}
