% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{listingw}
\alias{listingw}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
listingw(
  path = ".",
  recurse = TRUE,
  show_hidden = FALSE,
  open = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
