% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_if_missing}
\alias{return_if_missing}
\title{If a value is missing, trigger the parent function's return}
\usage{
return_if_missing(x, value)
}
\arguments{
\item{x}{Focal input.}

\item{value}{If \code{x} is missing, \code{\link{return}} this input
from the parent function.}
}
\value{
If \code{x} is not missing, \code{NULL} is returned invisibly. If 
 \code{x} is missing, if \code{value} is provided, the result of 
 \code{\link{return}} with \code{value} as its input evaluated within the 
  parent function's environment is returned, otherwise \code{NULL} is 
  returned invisibly.
}
\description{
If the focal input is \code{NULL}, return \code{value} from
 the parent function. Should only be used within a function.
}
\examples{
 ff <- function(x = 1, missing_return = "hello"){
   return_if_missing(x, missing_return)
   x
 }
 ff()

}
