% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directree.R
\name{tree.create}
\alias{tree.create}
\title{Create a Directory Tree}
\usage{
tree.create(
  tree = list(),
  fsep = .Platform$file.sep,
  winslash = "\\\\",
  mustWork = NA,
  showWarnings = TRUE,
  recursive = TRUE,
  mode = "0777"
)
}
\arguments{
\item{tree}{\code{list} object describing a directory tree hierarchicaly as
ingested by \code{\link{tree_paths}}. See \code{Examples}.}

\item{fsep}{Path separator to use (assumed to be ASCII).
See \code{\link[base]{file.path}}.}

\item{winslash}{Separator to be used on Windows - ignored elsewhere. 
See \code{\link[base]{normalizePath}}.}

\item{mustWork}{\code{logical}: if \code{TRUE} then an error is given if 
the result cannot be determined; if \code{NA}, then a warning.
See \code{\link[base]{normalizePath}}.
\cr \cr
\strong{NB:} The default argument is changed to \code{FALSE} compared to
\code{NA}, given that the folders are being created and so are by 
default, assumed to not yet exist. The argument is retained, however,
as it can be useful for avoiding overwriting of existing folders.}

\item{showWarnings}{\code{logical}: should the warnings on failure be 
shown? 
See \code{\link[base]{dir.create}}.}

\item{recursive}{\code{logical}: should elements of the path other than 
the last be created? If \code{TRUE}, like the Unix command 
\code{mkdir -p}. 
\cr \cr
\strong{NB:} The default argument is changed to \code{TRUE} compared
to \code{\link[base]{dir.create}} and \code{\link{dirs.create}} in order
to create the full tree. Setting \code{recursive = FALSE} will generate
a "tree" with only its final (branch tip) folders created.}

\item{mode}{The mode to be used on Unix-alikes: it will be coerced by 
\code{\link[base]{as.octmode}}. For \code{\link[base]{Sys.chmod}} it 
is recycled along paths.
See \code{\link[base]{dir.create}}.}
}
\value{
\code{logical} vector of results of \code{\link[base]{dir.create}}
 on each of the \code{paths} (which name the vector).
 See \code{\link[base]{dir.create}}.
}
\description{
Creates a tree of directories with flexible structure by
 running \code{\link[base]{dirs.create}} over \code{paths} defined
 by the generalized \code{tree} input list, which is processed through
 \code{\link{tree_paths}}.
}
\examples{
 \dontrun{

   1
 
  }

}
