% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articulate.R, R/tree.R
\name{tree}
\alias{tree}
\alias{articulate}
\alias{articulate.list}
\title{Articulate input}
\usage{
articulate(x)

\method{articulate}{list}(x)

tree(paths = NULL)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}

\item{tree}{\code{list} defining the directory.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
Directory tree as a \code{character} \code{vector}.

Directory tree as a list.
}
\description{
ffff

ffff
}
\examples{
 \donttest{
  articulate()
 }

 \donttest{
  tree(".")
 }

}
