% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dir_package.R
\name{use_dir_package}
\alias{use_dir_package}
\title{Setup the package to use 'dir'}
\usage{
use_dir_package(..., recursive = TRUE, patch = FALSE, add_overrides = !patch)
}
\arguments{
\item{...}{Folders to add. You might provide strings, vectors or lists, they will be flattened into a vector.}

\item{recursive}{Boolean. Whether to add folders under the added folders recursively.}

\item{patch}{Boolean. Whether to patch some 'usethis' and 'devtools' functions so they
work with 'dir'. Recommended but since it is invasive it defaults to \code{FALSE}.
Forwarded to \code{dir::add()}}

\item{add_overrides}{Boolean. Whether to add overrides for 'usethis' and 'devtools'
functions in a \code{dir-overrides} environment on the search path. These can be
adjusted to your taste in the ".Rprofile".}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
This is the only function you need to use 'dir'. It sets up your local
".Rprofile" so you can use nested folders in your project. You can tweak the
".RProfile" manually if needed or re-rerun the function to override the setup code.
}
