% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{usethis}
\alias{usethis}
\alias{use_r}
\alias{use_test}
\title{'usethis' overrides}
\usage{
use_r(
  name = NULL,
  open = rlang::is_interactive(),
  test_structure = getOption("dir.test_structure", "nested")
)

use_test(
  name = NULL,
  open = rlang::is_interactive(),
  test_structure = getOption("dir.test_structure", "nested")
)
}
\arguments{
\item{name}{Either a string giving a file name (without directory) or
\code{NULL} to take the name from the currently open file in RStudio.}

\item{open}{Whether to open the file for interactive editing.}

\item{test_structure}{Either "flat" or "nested". If "flat" we'll replace
"/" with "--" in paths to have flat test names. If "nested" we'll reproduce
the nested structure. In tests we consider folders at the top level and
sub folder of the R folder at the same level, if
you name folders in R an at the top level the same there will be ambiguity..}
}
\value{
These functions return \code{NULL} invisibly, they're called for side effects.
}
\description{
Wrappers around 'usethis' functions that behave the same but support a flexible
folder structure. Not needed if you use \code{use_dir_package(, patch = TRUE)},
which will patch the actual 'usethis' functions directly.
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}}
}
