% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Verify that folders exist.}
\usage{
verify(paths = NULL)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}
}
\value{
Throws an error if any of the folders do not exist, otherwise \code{NULL}
}
\description{
Throws an error if any of the folders (specified by \code{path}) do 
 not exist.
}
\examples{
 \donttest{
  verify(".")
 }

}
