% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add Folders to a R package}
\usage{
add(..., recursive = TRUE, patch = FALSE)
}
\arguments{
\item{...}{Folders to add. You might provide strings, vectors or lists, they will be flattened into a vector.}

\item{recursive}{Boolean. Whether to add folders under the added folders recursively.}

\item{patch}{Boolean. Whether to patch some 'usethis' and 'devtools' functions so they
work with 'dir'. Recommended but since it is invasive it defaults to \code{FALSE}.
Forwarded to \code{dir::add()}}
}
\description{
\code{add()} makes sure scripts stored outside of the "R" folder are added to \code{.Rbuildignore}
and turns these scripts into objects stored in \code{R/sysdata.rda}.
\code{use_dir_package()} will place a call to \code{dir::add()} in your ".RProfile".
You shouldn't need to call it manually.
}
