% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{covr}
\alias{covr}
\alias{report}
\alias{package_coverage}
\title{'covr' overrides}
\usage{
report(
  x = dir::package_coverage(),
  file = file.path(tempdir(), paste0(attr(x, "package")$package \%||\% "coverage",
    "-report.html")),
  browse = interactive()
)

package_coverage(
  path = ".",
  type = c("tests", "vignettes", "examples", "all", "none"),
  combine_types = TRUE,
  relative_path = TRUE,
  quiet = TRUE,
  clean = TRUE,
  line_exclusions = NULL,
  function_exclusions = NULL,
  code = character(),
  install_path = normalizePath(tempfile("R_LIBS"), mustWork = FALSE),
  ...,
  exclusions,
  pre_clean = TRUE
)
}
\arguments{
\item{x}{a coverage dataset, defaults to running \code{package_coverage()}.}

\item{file}{The report filename.}

\item{browse}{whether to open a browser to view the report.}

\item{path}{file path to the package.}

\item{type}{run the package \sQuote{tests}, \sQuote{vignettes},
\sQuote{examples}, \sQuote{all}, or \sQuote{none}. The default is
\sQuote{tests}.}

\item{combine_types}{If \code{TRUE} (the default) the coverage for all types
is simply summed into one coverage object. If \code{FALSE} separate objects
are used for each type of coverage.}

\item{relative_path}{whether to output the paths as relative or absolute
paths. If a string, it is interpreted as a root path and all paths will be
relative to that root.}

\item{quiet}{whether to load and compile the package quietly, useful for
debugging errors.}

\item{clean}{whether to clean temporary output files after running, mainly
useful for debugging errors.}

\item{line_exclusions}{a named list of files with the lines to exclude from
each file.}

\item{function_exclusions}{a vector of regular expressions matching function
names to exclude. Example \verb{print\\\\\\.} to match print methods.}

\item{code}{A character vector of additional test code to run.}

\item{install_path}{The path the instrumented package will be installed to
and tests run in. By default it is a path in the R sessions temporary
directory. It can sometimes be useful to set this (along with \code{clean = FALSE}) to help debug test failures.}

\item{...}{Additional arguments passed to \code{\link[tools:testInstalledPackage]{tools::testInstalledPackage()}}.}

\item{exclusions}{\sQuote{Deprecated}, please use \sQuote{line_exclusions} instead.}

\item{pre_clean}{whether to delete all objects present in the src directory before recompiling}
}
\value{
These functions return \code{NULL} invisibly, they're called for side effects.
}
\description{
Wrappers around 'covr' functions that behave the same but support a flexible
folder structure. Not needed if you use \code{use_dir_package(, patch = TRUE)},
which will patch the actual 'covr' functions directly.
}
\seealso{
\code{\link[devtools:document]{devtools::document()}}, \code{\link[devtools:check]{devtools::check()}}, \code{\link[devtools:build]{devtools::build()}}
}
