% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir-package.R
\docType{package}
\name{dir-package}
\alias{dir}
\alias{dir-package}
\title{dir: Make HTML directory listing}
\description{
Tools to make it easier to make a pretty directory listing with collapsible folders.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://emitanaka.org/dir}
  \item \url{https://github.com/emitanaka/dir}
  \item Report bugs at \url{https://github.com/emitanaka/dir/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
